sequence-import/requirements.txt
requests==2.31.0
pandas==1.5.3
rich==13.7.1
python-dotenv==1.0.1 